// src/lib/types/generated/SkillConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConfigSource } from "./ConfigSource";

export type SkillConfig = { name: string, description: string | null, file_path: string, source: ConfigSource, argument_hint: string | null, allowed_tools: Array<string> | null, disable_model_invocation: boolean | null, author: string | null, version: string | null, category: string | null, content: string, };