// src/lib/types/generated/HookConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConfigSource } from "./ConfigSource";

export type HookConfig = { filename: string, file_path: string, source: ConfigSource, event_type: string, matcher: string | null, timeout: number | null, description: string | null, };