// src/lib/types/generated/MemoryConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConfigSource } from "./ConfigSource";
import type { ProviderKind } from "./ProviderKind";

export type MemoryConfig = { name: string, description: string | null, file_path: string, source: ConfigSource, provider: ProviderKind, memory_type: string | null, content: string, deprecated: boolean, };