// src/lib/types/generated/RuleConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConfigSource } from "./ConfigSource";
import type { ProviderKind } from "./ProviderKind";

export type RuleConfig = { filename: string, file_path: string, source: ConfigSource, provider: ProviderKind, language: string | null, description: string | null, always_apply: boolean | null, globs: Array<string> | null, content: string, deprecated: boolean, };